feat(settings): move command-line tool to a new General tab
The `termio` command-line tool is an app integration (install the binary to
PATH), not an agent feature, so it doesn't belong in the Agents tab. Add a
General tab (first in the sidebar) and move it there; the Agents tab's
Orchestration section now carries only Session control.
